requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-APScheduler==1.13.1
Flask-Caching==2.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
python-dotenv==1.0.1
redis==5.0.1
requests==2.32.3
SQLAlchemy==2.0.28
todoist_api_python==2.1.7
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
